fulstack/group_backend/requirements.txt
Flask>=3.1.0
Flask-SQLAlchemy>=3.1.0
Flask-Login>=0.6.0
Flask-JWT-Extended>=4.4.0
Flask-Migrate>=4.0.0
python-dotenv>=1.0.0